refactor: extracting render model calculations 2
